+++ migrations/1669623288983-add-manage-custom-notifications-permission.ts
@@ -10,6 +10,14 @@ import { startDatabaseForMigration } from '../src/utils/migrations/database.help
 export const up = async () => {
   await startDatabaseForMigration();
   const type = 'can_manage_custom_notifications';
+
+  // check if permission already exists
+  const permissionExists = await Permission.exists({ type, global: false });
+  if (permissionExists) {
+    logger.info(`${type} permission already exists`);
+    return;
+  }
+
   const permission = new Permission({
     type,
     global: false,