+++ src/routes/download/index.ts
@@ -148,15 +148,15 @@ router.get('/form/records/:id/history', async (req, res) => {
     let filters: {
       fromDate?: Date;
       toDate?: Date;
-      field?: string;
+      fields?: string[];
     } = {};
     if (req.query) {
-      const { from, to, field } = req.query as any;
+      const { from, to, fields } = req.query as any;
       filters = Object.assign(
         {},
         from === 'NaN' ? null : { fromDate: new Date(parseInt(from, 10)) },
         to === 'NaN' ? null : { toDate: new Date(parseInt(to, 10)) },
-        !field ? null : { field }
+        !fields ? null : { fields: fields.split(',') }
       );
 
       if (filters.toDate) filters.toDate.setDate(filters.toDate.getDate() + 1);
@@ -189,7 +189,7 @@ router.get('/form/records/:id/history', async (req, res) => {
       const meta: RecordHistoryMeta = {
         form: form.name,
         record: record.incrementalId,
-        field: filters.field || req.t('history.allFields'),
+        fields: filters.fields?.join(',') || '',
         fromDate: filters.fromDate
           ? filters.fromDate.toLocaleDateString(dateLocale)
           : '',
@@ -205,6 +205,7 @@ router.get('/form/records/:id/history', async (req, res) => {
           ability,
         }
       ).getHistory();
+      const fields = filters.fields;
       const history = unfilteredHistory
         .filter((version) => {
           let isInDateRange = true;
@@ -216,16 +217,16 @@ router.get('/form/records/:id/history', async (req, res) => {
 
           // filtering by field
           const changesField =
-            !filters.field ||
-            !!version.changes.find((item) => item.field === filters.field);
+            !fields ||
+            !!version.changes.find((item) => fields.includes(item.field));
 
           return isInDateRange && changesField;
         })
         .map((version) => {
           // filter by field for each verison
-          if (filters.field) {
-            version.changes = version.changes.filter(
-              (change) => change.field === filters.field
+          if (fields) {
+            version.changes = version.changes.filter((change) =>
+              fields.includes(change.field)
             );
           }
           return version;
@@ -399,12 +400,24 @@ router.get('/invite', async (req, res) => {
 });
 
 /**
- * Export all users of the platform
+ * Export selected users of the platform,
+ * if a list with ids is not provided in the body, export all of them
  */
-router.get('/users', async (req, res) => {
+router.post('/users', async (req, res) => {
   const ability: AppAbility = req.context.user.ability;
   if (ability.can('read', 'User')) {
-    const users: any[] = await User.find({}).populate({
+    const ids =
+      req.body.users?.map((id) => new mongoose.Types.ObjectId(id)) || [];
+
+    const filters = {};
+
+    // If ids are provided, filter the users
+    if (ids && ids.length > 0)
+      Object.assign(filters, {
+        _id: { $in: ids },
+      });
+
+    const users: any[] = await User.find(filters).populate({
       path: 'roles',
       match: { application: { $eq: null } },
     });
@@ -414,12 +427,16 @@ router.get('/users', async (req, res) => {
 });
 
 /**
- * Export the users of a specific application
+ * Export the users of a specific application,
+ * if a list with ids is not provided in the body, export all of them
  */
-router.get('/application/:id/users', async (req, res) => {
+router.post('/application/:id/users', async (req, res) => {
   const ability: AppAbility = req.context.user.ability;
   if (ability.can('read', 'User')) {
-    const aggregations = [
+    const ids =
+      req.body.users?.map((id) => new mongoose.Types.ObjectId(id)) || [];
+
+    const aggregations: any[] = [
       // Left join
       {
         $lookup: {
@@ -449,6 +466,11 @@ router.get('/application/:id/users', async (req, res) => {
       // Filter users that have at least one role in the application.
       { $match: { 'roles.0': { $exists: true } } },
     ];
+
+    if (ids.length > 0)
+      // Filter users that are in the list of ids.
+      aggregations.push({ $match: { _id: { $in: ids } } });
+
     const users = await User.aggregate(aggregations);
     return buildUserExport(req, res, users);
   }