+++ src/schema/mutation/duplicatePage.mutation.ts
@@ -1,7 +1,7 @@
 import { GraphQLID, GraphQLNonNull, GraphQLError } from 'graphql';
 import { AppAbility } from '@security/defineUserAbility';
 import { PageType } from '../types';
-import { Application, Page, Role } from '@models';
+import { Application, Page, Role, Step, Workflow } from '@models';
 import { duplicatePage } from '../../services/page.service';
 
 /**
@@ -12,45 +12,104 @@ import { duplicatePage } from '../../services/page.service';
 export default {
   type: PageType,
   args: {
-    id: { type: new GraphQLNonNull(GraphQLID) },
+    page: { type: GraphQLID },
+    step: { type: GraphQLID },
     application: { type: new GraphQLNonNull(GraphQLID) },
   },
   async resolve(parent, args, context) {
+    // Check ability
     const user = context.user;
     if (!user) {
       throw new GraphQLError(context.i18next.t('common.errors.userNotLogged'));
     }
     const ability: AppAbility = user.ability;
+    // Check parameters
+    if (!args.page && !args.step) {
+      throw new GraphQLError(
+        context.i18next.t(
+          'mutations.application.content.duplicate.errors.missingArguments'
+        )
+      );
+    }
+    if (args.page && args.step) {
+      throw new GraphQLError(
+        context.i18next.t(
+          'mutations.application.content.duplicate.errors.incompatibleArguments'
+        )
+      );
+    }
+    // Find application
     const application = await Application.findById(args.application);
     if (!application)
       throw new GraphQLError(context.i18next.t('common.errors.dataNotFound'));
-
     // Check access to the application
     if (ability.can('update', application)) {
-      const page = await Page.findById(args.id);
-      const pageApplication = await Application.findOne({
-        pages: { $in: [args.id] },
-      });
-      if (ability.can('update', pageApplication)) {
-        // Get list of roles for default permissions
-        const roles = await Role.find({ application: application._id });
-        const permissions = {
-          canSee: roles.map((x) => x.id),
-          canUpdate: [],
-          canDelete: [],
-        };
-        // Create the new page and set the permissions
-        const newPage = await duplicatePage(
-          page,
-          `${page.name} Copy`,
-          permissions
-        );
-        const update = {
-          //modifiedAt: new Date(),
-          $push: { pages: newPage.id },
-        };
-        await Application.findByIdAndUpdate(args.application, update);
-        return newPage;
+      if (args.page) {
+        // If page
+        const page = await Page.findById(args.page);
+        if (!page) throw new GraphQLError('common.errors.dataNotFound');
+        const pageApplication = await Application.findOne({
+          pages: { $in: [args.page] },
+        });
+        if (!pageApplication)
+          throw new GraphQLError('common.errors.dataNotFound');
+        if (ability.can('update', pageApplication)) {
+          // Get list of roles for default permissions
+          const roles = await Role.find({ application: application._id });
+          const permissions = {
+            canSee: roles.map((x) => x.id),
+            canUpdate: [],
+            canDelete: [],
+          };
+          // Create the new page and set the permissions
+          const newPage = await duplicatePage(
+            page,
+            `${page.name} Copy`,
+            permissions
+          );
+          const update = {
+            $push: { pages: newPage.id },
+          };
+          await Application.findByIdAndUpdate(args.application, update);
+          return newPage;
+        }
+      } else {
+        // If step
+        const step = await Step.findById(args.step);
+        if (!step) throw new GraphQLError('common.errors.dataNotFound');
+        const workflow = await Workflow.findOne({
+          steps: { $in: [args.step] },
+        });
+        if (!workflow) throw new GraphQLError('common.errors.dataNotFound');
+        const page = await Page.findOne({
+          content: workflow._id,
+        });
+        if (!page) throw new GraphQLError('common.errors.dataNotFound');
+        const stepApplication = await Application.findOne({
+          pages: { $in: [page._id] },
+        });
+        if (!stepApplication)
+          throw new GraphQLError('common.errors.dataNotFound');
+        if (ability.can('update', stepApplication)) {
+          // Get list of roles for default permissions
+          const roles = await Role.find({ application: application._id });
+          const permissions = {
+            canSee: roles.map((x) => x.id),
+            canUpdate: [],
+            canDelete: [],
+          };
+          // Create the new page and set the permissions
+          const newPage = await duplicatePage(
+            new Page(step),
+            `${step.name} Copy`,
+            permissions
+          );
+          const update = {
+            $push: { pages: newPage.id },
+          };
+          await Application.findByIdAndUpdate(args.application, update);
+          return newPage;
+        }
       }
     } else {
       throw new GraphQLError(